Return float in median func

+++ statistics.py
@@ -18,8 +18,8 @@ def median(values):
     """Return the median value of the values list."""
     sorted_values = sorted(values)
     if len(sorted_values) % 2 == 0:
-        return (sorted_values[int(len(sorted_values) / 2) - 1] + sorted_values[int(len(sorted_values) / 2)]) / 2
-    return sorted_values[int(len(sorted_values) / 2)]
+        return float((sorted_values[int(len(sorted_values) / 2) - 1] + sorted_values[int(len(sorted_values) / 2)]) / 2)
+    return float(sorted_values[int(len(sorted_values) / 2)])
 
 
 def population_statistics(feature_description, data, treatment, target, threshold, is_above, statistic_functions):
